+++ apps/web/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @vegastack/pages-web
 
+## 0.2.0-next.12
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vegastack/pages-core@0.2.0-next.12
+  - @vegastack/pages-mcp@0.2.0-next.12
+  - @vegastack/pages-renderer@0.2.0-next.12
+  - @vegastack/pages-services@0.2.0-next.12
+  - @vegastack/pages-ui@0.2.0-next.12
+
 ## 0.2.0-next.11
 
 ### Patch Changes