+++ packages/services/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @vegastack/pages-services
 
+## 0.2.0-next.12
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vegastack/pages-core@0.2.0-next.12
+  - @vegastack/pages-db@0.2.0-next.12
+  - @vegastack/pages-renderer@0.2.0-next.12
+
 ## 0.2.0-next.11
 
 ### Patch Changes